+++ resources/views/account.blade.php
@@ -33,7 +33,8 @@
                         <div class="alert alert-dark mt-5" role="alert">
                             <h4 class="alert-heading">Delete started!</h4>
                             <p>A job has been added to delete your account, we should be done in a minute or two!</p>
-                            <p class="mb-0">You can check the status by refreshing your browser, when the job is complete you will be signed-out.</p>
+                            <p>You have been logged out, if you refresh you will be back at the login screen.</p>
+                            <p>You will get an email when your account has been deleted.</p>
                         </div>
                     @endif
                 @endif